requirement.txt
flask
psycopg2-binary==2.9.6
sqlalchemy
click
flask
python-dotenv
